pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ops-proxy"
version = "1.0.0"
description = "OpsProxy daemon for controlled network access"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "watchdog>=3.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
ops-proxy = "ops_proxy.cli:main"

[tool.setuptools.packages.find]
where = ["."]
